[feat]: Add Select Time Fragment
