+++ core/designsystem/src/main/res/values/strings.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="all_hour">시간</string>
+    <string name="use_time_goal_title">선택한 앱의 목표 이용 시간을\n설정해 주세요</string>
+    <string name="use_time_goal_description">목표 이용 시간은 최대 1시간 59분까지\n설정할 수 있어요</string>
+    <string name="all_minute">분</string>
+</resources>
