+++ feature/challenge/src/main/java/com/hmh/hamyeonham/challenge/appadd/AppSelectionViewModel.kt
@@ -1,0 +1,9 @@
+package com.hmh.hamyeonham.challenge.appadd
+
+import androidx.lifecycle.ViewModel
+
+class AppSelectionViewModel : ViewModel() {
+    val selectedApp: MutableList<String> = mutableListOf()
+
+    fun nonSelected(): Boolean = selectedApp.isEmpty()
+}
