+++ feature/challenge/src/main/java/com/hmh/hamyeonham/challenge/appadd/time/SetGoalTimeFragment.kt
@@ -1,4 +1,4 @@
-package com.hmh.hamyeonham.challenge.appselection.fragment
+package com.hmh.hamyeonham.challenge.appadd.time
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,16 +6,16 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.hmh.hamyeonham.common.view.viewBinding
-import com.hmh.hamyeonham.feature.challenge.databinding.FrargmentAppSelectionBinding
+import com.hmh.hamyeonham.feature.challenge.databinding.FragmentSetGoalTimeBinding
 
-class AppSelectionFragment:Fragment() {
-    private val binding by viewBinding(FrargmentAppSelectionBinding::bind)
+class SetGoalTimeFragment : Fragment() {
+    private val binding by viewBinding(FragmentSetGoalTimeBinding::bind)
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        return FrargmentAppSelectionBinding.inflate(inflater,container,false).root
+        return FragmentSetGoalTimeBinding.inflate(inflater, container, false).root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
