--- feature/challenge/src/main/java/com/hmh/hamyeonham/challenge/appselection/AppSelectionViewModel.kt
@@ -1,9 +1,0 @@
-package com.hmh.hamyeonham.challenge.appselection
-
-import androidx.lifecycle.ViewModel
-
-class AppSelectionViewModel : ViewModel() {
-    val selectedApp: MutableList<String> = mutableListOf()
-
-    fun nonSelected(): Boolean = selectedApp.isEmpty()
-}
