+++ feature/challenge/src/main/res/layout/fragment_set_goal_time.xml
@@ -11,7 +11,7 @@
         android:layout_marginHorizontal="21dp"
         android:layout_marginTop="30dp"
         android:lineSpacingExtra="6sp"
-        android:text="@string/onboarding_screentime_goal_title"
+        android:text="@string/use_time_goal_title"
         android:textColor="?white_text"
         android:textSize="21sp"
         android:textStyle="bold"
@@ -25,7 +25,7 @@
         android:layout_height="wrap_content"
         android:layout_marginStart="21dp"
         android:layout_marginTop="7dp"
-        android:text="@string/onboarding_screentime_goal_description"
+        android:text="@string/use_time_goal_description"
         android:textColor="?gray2"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/tv_onboarding_screentime_goal_title" />
