+++ feature/onboarding/build.gradle.kts
@@ -10,5 +10,4 @@ android {
 dependencies {
     implementation(projects.core.common)
     implementation(projects.core.designsystem)
-    implementation(project(":feature:main"))
 }
