+++ feature/onboarding/src/main/res/layout/fragment_on_boarding_select_use_time.xml
@@ -13,7 +13,7 @@
         android:layout_height="0dp"
         android:layout_marginHorizontal="21dp"
         android:layout_marginTop="30dp"
-        android:text="@string/onboarding_use_time_goal_title"
+        android:text="@string/use_time_goal_title"
         android:textColor="?white_text"
         android:textStyle="bold"
         android:textSize="21sp"
@@ -26,7 +26,7 @@
         android:layout_height="0dp"
         android:layout_marginStart="21dp"
         android:layout_marginTop="7dp"
-        android:text="@string/onboarding_use_time_goal_description"
+        android:text="@string/use_time_goal_description"
         android:textColor="?gray2"
         style="@style/TextAppearance.Hmh.BodyMedium"
         app:layout_constraintStart_toStartOf="parent"
