+++ feature/onboarding/src/main/res/values/strings.xml
@@ -18,10 +18,7 @@
     <string name="onboarding_select_app_description">목표 이용 시간이 지나면 앱이 잠겨요\n선택하고 싶은 앱은 언제든지 추가할 수 있어요</string>
     <string name="onboarding_screentime_goal_description">목표 이용 시간은 최대 6시간까지 설정할 수 있어요</string>
     <string name="onboarding_screentime_goal_title">총 목표 이용 시간을\n설정해 주세요</string>
-    <string name="all_hour">시간</string>
-    <string name="onboarding_use_time_goal_title">선택한 앱의 목표 이용 시간을\n설정해 주세요</string>
-    <string name="onboarding_use_time_goal_description">목표 이용 시간은 최대 1시간 59분까지\n설정할 수 있어요</string>
-    <string name="all_minute">분</string>
+
     <string name="onboarding_done_sing_up_image">celebrate sign up image</string>
     <string name="onboarding_done_sing_up_title">회원가입 완료</string>
     <string name="onboarding_done_sing_up_description">이제 하면함에서 블랙홀 탈출을 위한\n여정을 시작해 볼까요?</string>
