+++ feature/onboarding/src/main/res/values/themes.xml
@@ -1,10 +1,3 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-
-    <style name="AppTheme.NumberPicker" parent="Base.Theme.HMH">
-        <item name="colorControlNormal">?gray6</item>
-        <item name="android:background">#00FFFFFF</item>
-        <item name="android:editTextColor">?white_text</item>
-        <item name ="android:textSize">24sp</item>
-    </style>
 </resources>
